# internal/models/userMod.go
package models

import "encoding/json"

type User struct {
	ID         int    `json:"id"`
	Username   string `json:"username"`
	FirstName  string `json:"firstName"`
	LastName   string `json:"lastName"`
	Email      string `json:"email"`
	Password   string `json:"password"`
	Phone      string `json:"phone"`
	UserStatus int    `json:"userStatus"`
}

func (u *User) ToJson() []byte {
	data, _ := json.Marshal(u)
	return data
}